feature: add search tool feature: escape and enter key interact with dialog windows
x

+++ src/main/java/com/d4alencar/crud/App.java
@@ -7,7 +7,7 @@ import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
 
-public class App extends JFrame {
+public class App extends JFrame{
 
   static JFrame frame;
   static String title = "CRUD BY d4alencar";
@@ -26,7 +26,7 @@ public class App extends JFrame {
     frame = new JFrame(title);
     frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
     frame.setSize(400, 300);
-    
+
     //PANELS
     CardLayout cardLayout = new CardLayout();
     JPanel mainPanel = new JPanel(cardLayout);
@@ -39,26 +39,39 @@ public class App extends JFrame {
 
 
     //MENU BAR PANEL
+    String[] options = {"title", "author", "year"};
+
     JButton addBookButton = new JButton("Add");
     JButton editBookButton = new JButton("Edit");
     JButton deleteBookButton = new JButton("Delete");
+    JTextField searchField = new JTextField();
+    JComboBox<String> searchOptions = new JComboBox<>(options);
+    JButton searchButton = new JButton("Search");
 
     menuBarPanel.add(addBookButton);
     menuBarPanel.add(editBookButton);
     menuBarPanel.add(deleteBookButton);
+    menuBarPanel.add(searchField);
+    menuBarPanel.add(searchOptions);
+    menuBarPanel.add(searchButton);
 
+    searchField.addKeyListener(new KeyAdapter() {
+      @Override
+      public void keyPressed(KeyEvent e) {
+        if(e.getKeyChar() == '\n') {
+          searchButton.doClick();
+        }
+      }
+    });
 
     //LIST PANEL
     listBookPanel.add(new JScrollPane(list), BorderLayout.CENTER);
     listBookPanel.add(menuBarPanel, BorderLayout.NORTH);
 
-
     mainPanel.add(listBookPanel, "list");
-    
 
-    //Buttons Actions
     addBookButton.addActionListener(e -> openAddDialog(frame, libraryService));
-    
+
     editBookButton.addActionListener(new ActionListener() {
       @Override
       public void actionPerformed(ActionEvent e) {
@@ -75,21 +88,55 @@ public class App extends JFrame {
       @Override
       public void actionPerformed (ActionEvent e) {
         int selectedIndex = list.getSelectedIndex();
-        
-        if(libraryService.deleteBook(selectedIndex)){
-          JOptionPane.showMessageDialog(frame, "book deleted");
+
+        if(selectedIndex != -1) {
+          System.out.println("test: "+selectedIndex);
+          String[] optionsTitle = {"Yes.", "No."};
+          int responseIndex = JOptionPane.showOptionDialog(frame, "Do you really want to delete this book?", "Deleting book", 2, 3, null, optionsTitle, 0);
+          System.out.println(responseIndex);
+          
+          if(responseIndex == 0) {
+            if(libraryService.deleteBook(selectedIndex)){
+              JOptionPane.showMessageDialog(frame, "book deleted");
+            } else {
+              JOptionPane.showMessageDialog(frame, "error deleting book");
+            }
+          }
         } else {
-          JOptionPane.showMessageDialog(frame, "error deleting book");
+          JOptionPane.showMessageDialog(frame, "Error deleting book: no index selected");
         }
       }
 
     });
+
+    searchButton.addActionListener(new ActionListener() {
+      @Override
+      public void actionPerformed (ActionEvent e) {
+        String option = searchOptions.getSelectedItem().toString();
+        String input = searchField.getText();
+        if(option.equals("year")) {
+          try {
+            if(!input.isEmpty()) {
+              Integer.parseInt(input);
+              libraryService.searchBook(input, option);
+            } else {
+              libraryService.searchBook(input, option);
+            }
+          } catch (NumberFormatException a) {
+            JOptionPane.showMessageDialog(frame, "input invalid for year search");
+          }
+        } else {
+          libraryService.searchBook(input, option);
+        }
+      }
+    });
+
     frame.add(mainPanel);
     frame.setVisible(true);
+    
   }
 
   private void openAddDialog(JFrame parent, LibraryService libraryService) {
- 
     JDialog dialog = new JDialog(parent, "Add Book", true);
     dialog.setSize(300, 150);
     dialog.setLocationRelativeTo(parent);
@@ -144,7 +191,6 @@ public class App extends JFrame {
     cons.anchor = GridBagConstraints.CENTER;
     panel.add(newBookPublishYear, cons);
 
-    //cons.insets = new Insets (3, 3, 3, 3);
     JButton closeDialogButton = new JButton("Close");
     cons.gridy = 3;
     cons.gridx = 0;
@@ -162,32 +208,57 @@ public class App extends JFrame {
     panel.add(addNewBookButton, cons);
 
     closeDialogButton.addActionListener(e -> dialog.dispose());
-    
+
     addNewBookButton.addActionListener (new ActionListener(){
       @Override
       public void actionPerformed (ActionEvent e) {
-        Book newBook = new Book(0, newBookTitle.getText(), Integer.parseInt(newBookPublishYear.getText()), newBookAuthor.getText());
-        if(libraryService.addBook(newBook)) {
-          JOptionPane.showMessageDialog(frame, "book added");
+        
+        String titleBuff = newBookTitle.getText();
+        String authorBuff = newBookAuthor.getText();
+        String yearBuff = newBookPublishYear.getText();
+        if(!titleBuff.isEmpty() && !authorBuff.isEmpty() && !yearBuff.isEmpty()){
+          try {
+            Book newBook = new Book(0, titleBuff, Integer.parseInt(yearBuff), authorBuff);
+            if(libraryService.addBook(newBook)) {
+              JOptionPane.showMessageDialog(frame, "book added");
+              dialog.dispose();
+            } else {
+              JOptionPane.showMessageDialog(frame, "error adding book");
+            }
+          } catch (NumberFormatException eN) {
+            JOptionPane.showMessageDialog(frame, "invalid caracter in year field");
+          }
         } else {
-          JOptionPane.showMessageDialog(frame, "error adding book");
+          JOptionPane.showMessageDialog(frame, "All fields must be completed.");
         }
+      }
+    });
+
+    JRootPane rootPane = dialog.getRootPane();
+    KeyStroke escapeKey = KeyStroke.getKeyStroke("ESCAPE");
+
+    rootPane.getInputMap(JComponent.WHEN_IN_FOCUSED_WINDOW).put(escapeKey, "closeDialog");
+    rootPane.getActionMap().put("closeDialog", new AbstractAction() {
+      @Override
+      public void actionPerformed(ActionEvent e) {
         dialog.dispose();
       }
     });
 
+    rootPane.setDefaultButton(addNewBookButton);
     dialog.add(panel);
     dialog.setVisible(true);
+
   }
 
     private void openEditDialog(JFrame parent, LibraryService libraryService, JList<Book> list, int id) {
-
 
       Book selectedBook = libraryService.model.get(id);
       
       JDialog dialog = new JDialog(parent, "Edit book", true);
       dialog.setSize(300, 150);
       dialog.setLocationRelativeTo(frame);
+
 
       JPanel panel = new JPanel(new GridBagLayout());
       GridBagConstraints cons = new GridBagConstraints();
@@ -258,27 +329,51 @@ public class App extends JFrame {
       titleField.setText(selectedBook.getTitle());
       authorField.setText(selectedBook.getAuthor());
       publishYearField.setText(String.valueOf(selectedBook.getYear()));
-
       closeDialogButton.addActionListener(e -> dialog.dispose());
 
       saveModButton.addActionListener(new ActionListener() {
         @Override
         public void actionPerformed(ActionEvent e) {
           int selectedIndex = list.getSelectedIndex();
-          Book selectedBook = libraryService.model.get(selectedIndex);
 
-          Book updatedBook = new Book(selectedBook.getId(), titleField.getText(), Integer.parseInt(publishYearField.getText()), authorField.getText());
+          String titleBuff = titleField.getText();
+          String yearBuff = publishYearField.getText();
+          String authorBuff = authorField.getText();
 
-          if(libraryService.editBook(selectedIndex, updatedBook)) {
-            JOptionPane.showMessageDialog(frame, "book edited");
+          if(!titleBuff.isEmpty() && !yearBuff.isEmpty() && !authorBuff.isEmpty()) {
+            
+            try {
+              Book updatedBook = new Book(selectedBook.getId(), titleBuff, Integer.parseInt(yearBuff), authorBuff);
+
+              if(libraryService.editBook(selectedIndex, updatedBook)) {
+                JOptionPane.showMessageDialog(frame, "book edited");
+                dialog.dispose();
+              } else {
+                JOptionPane.showMessageDialog(frame, "Equals");
+              }
+            } catch (NumberFormatException nf) {
+              JOptionPane.showMessageDialog(frame, "Invalid caracter in year field");
+            }
           } else {
-            JOptionPane.showMessageDialog(frame, "error editing book");
+            JOptionPane.showMessageDialog(frame, "All fields must be filled.");
           }
+        }
+      });
+
+      JRootPane rootPane = dialog.getRootPane();
+      KeyStroke escapeKey = KeyStroke.getKeyStroke("ESCAPE");
+
+      rootPane.getInputMap(JComponent.WHEN_IN_FOCUSED_WINDOW).put(escapeKey, "closeDialog");
+      rootPane.getActionMap().put("closeDialog", new AbstractAction() {
+        @Override
+        public void actionPerformed(ActionEvent e) {
           dialog.dispose();
         }
       });
 
+      rootPane.setDefaultButton(saveModButton);
+
       dialog.add(panel);
       dialog.setVisible(true);
-    }
+  }
 }
